+++ examples/html/vite.config.ts
@@ -2,6 +2,9 @@
 import { defineConfig } from 'vitest/config';
 import reactPlugin from '@vitejs/plugin-react';
 import vuePlugin from '@vitejs/plugin-vue';
+import htmlTransform from '@web-widget/html/transform';
+import reactTransform from '@web-widget/react/transform';
+import vueTransform from '@web-widget/vue/transform';
 import { webRouterPlugin, webWidgetPlugin } from '@web-widget/vite-plugin';
 import { htmlCompress } from '@web-widget/html/vite-plugin';
 
@@ -22,7 +25,7 @@ export default defineConfig({
     reactPlugin(),
     vuePlugin(),
     webWidgetPlugin({
-      adapters: ['@web-widget/html', '@web-widget/react', '@web-widget/vue'],
+      transforms: [htmlTransform, reactTransform, vueTransform],
     }),
   ],
   build: {