+++ packages/preact/src/transform.ts
@@ -1,0 +1,7 @@
+import type { WidgetTransform } from '@web-widget/schema';
+
+export default {
+  name: 'preact',
+  extensions: ['.tsx', '.jsx'],
+  adapter: '@web-widget/preact/adapter',
+} satisfies WidgetTransform;