+++ packages/preact/tsup.config.ts
@@ -3,6 +3,7 @@ export default defineConfig({
   dts: true,
   entry: {
     preact: 'src/index.ts',
+    'preact.transform': 'src/transform.ts',
     'preact.adapter.server': 'src/adapter.server.ts',
     'preact.adapter.client': 'src/adapter.client.ts',
   },