+++ packages/svelte/tsup.config.ts
@@ -3,6 +3,7 @@ export default defineConfig({
   dts: true,
   entry: {
     svelte: 'src/index.ts',
+    'svelte.transform': 'src/transform.ts',
     'svelte.adapter.server': 'src/adapter.server.ts',
     'svelte.adapter.client': 'src/adapter.client.ts',
   },